split essentials in index/preset
